Working with font system input text because it was broken due to font system change and trying to fix vulkan errors

+++ Renderer/Source/Texture.h
@@ -21,7 +21,7 @@ enum class TextureCreateFlagBits {
 };
 struct TextureCoords {
 	Float2 Coords[4];
-	uint64_t Width, Height{};
+	uint32_t Width, Height{};
 
 };
 struct TextureAtlasData {
